.history/requirements_20260307150151.txt
streamlit
pandas
numpy
plotly
seaborn
matplotlib
openpyxl
openai
scikit-learn
networkx
python-telegram-bot
reportlab
streamlit-aggrid==0.3.4.post3
kaleido